finished connecting database to api, dealing with user conflicts on database
