+++ database.conf
@@ -1,5 +1,5 @@
-POSTGRES_USER=user1
+POSTGRES_USER=postgres
 POSTGRES_PASSWORD=password
 POSTGRES_HOST=postgres-database
-POSTGRES_PORT=5002
-POSTGRES_DB=nqueens
+POSTGRES_PORT=5432
+POSTGRES_DB=nQueens
