+++ results-api/Dockerfile
@@ -12,8 +12,8 @@ RUN pip install -r /tmp/requirements.txt
 WORKDIR /app   
 
 # Expose flask's port
-EXPOSE 5000
-# in docker container this will run on port 5000 with assigned ip address (reults-api)
+EXPOSE 5432
+# in docker container this will run on port 5432 with assigned ip address (reults-api)
 
 # Run the server
 CMD python server.py
